+++ src/packages/piral-cli/src/common/package.ts
@@ -300,6 +300,12 @@ export async function retrievePiletsInfo(entryFile: string) {
   };
 }
 
+function isValidDependency(name: string) {
+  // super simple check at the moment
+  // just to filter out things like "redux-saga/effects"
+  return name.indexOf('/') === -1 || name.indexOf('@') === 0;
+}
+
 export async function patchPiletPackage(
   root: string,
   name: string,
@@ -341,7 +347,7 @@ export async function patchPiletPackage(
       deps[name] = getDependencyVersion(name, info.devDependencies, piralDependencies);
       return deps;
     }, {}),
-    ...allExternals.reduce((deps, name) => {
+    ...allExternals.filter(isValidDependency).reduce((deps, name) => {
       deps[name] = piralDependencies[name] || 'latest';
       return deps;
     }, {}),